Update useTableau Hook
Pass args.version to the apiUrl and change the function call to use the minified version.

Both changes are to better reflect the readme:
"The version of the Tableau Embedding API to use. Defaults to "latest" which may change if you use Tableau Server or Cloud. Will always minify. If using a custom version[...]"